add: render feature and test code(main)
change: tab size 4 -> 2

+++ cyclone/src/integrator.rs
@@ -1,3 +1,3 @@
 pub trait Integratable {
-    fn integrate(&mut self, duration: f64);
+  fn integrate(&mut self, duration: f32);
 }
